add decorators

+++ src/server/common/helpers/dirUtils.ts
@@ -1,34 +1,29 @@
-import {F_OK} from 'constants';
 import fs from 'fs';
+import mkdirp from 'mkdirp';
+
+import {
+  fileExistsSync,
+  fileExistsAsync,
+} from './fileUtils';
 
 export function removeDirIfExistsSync(path: string) {
-  try {
-    fs.accessSync(path, F_OK);
+  if (fileExistsSync(path))
     fs.rmSync(path, {recursive: true});
-  } catch (e) {
-    if (!e || e.code !== 'ENOENT')
-      throw e;
-  }
 }
 
 export async function removeDirIfExistsAsync(path: string) {
-  try {
-    await fs.promises.access(path, F_OK);
+  if (await fileExistsAsync(path))
     await fs.promises.rm(path, {recursive: true});
-  } catch (e) {
-    if (!e || e.code !== 'ENOENT')
-      throw e;
-  }
 }
 
 export async function removeAndCreateDirSync(path: string) {
   removeDirIfExistsSync(path);
-  fs.mkdirSync(path);
+  mkdirp.sync(path);
   return path;
 }
 
 export async function removeAndCreateDirAsync(path: string) {
   await removeDirIfExistsAsync(path);
-  await fs.promises.mkdir(path);
+  await mkdirp(path);
   return path;
 }
